Struct rumqtt::MqttClient

pub struct MqttClient {
    pub opts: MqttOptions,
    pub state: MqttState,
    pub last_pkid: PacketIdentifier,
    pub pub0_channel_pending: u32,
    pub pub1_channel_pending: u32,
    pub pub2_channel_pending: u32,
    pub pub0_rx: Option<Receiver<Message>>,
    pub pub1_rx: Option<Receiver<Message>>,
    pub pub2_rx: Option<Receiver<Message>>,
    pub sub_rx: Option<Receiver<Vec<(TopicFilter, QualityOfService)>>>,
    pub nw_request_tx: Option<Sender<NetworkRequest>>,
    pub nw_notification_rx: Option<Receiver<NetworkNotification>>,
    pub misc_rx: Option<Receiver<MiscNwRequest>>,
    pub message_callback: Option<Arc<Box<Fn(Message) + Send + Sync>>>,
    pub publish_callback: Option<Arc<Box<Fn(Message) + Send + Sync>>>,
    pub poll: Poll,
}

Handles commands from Publisher and Subscriber. Saves MQTT state and takes care of retransmissions.

Fields
opts: MqttOptions state: MqttState last_pkid: PacketIdentifier pub0_channel_pending: u32 pub1_channel_pending: u32 pub2_channel_pending: u32 pub0_rx: Option<Receiver<Message>> pub1_rx: Option<Receiver<Message>> pub2_rx: Option<Receiver<Message>> sub_rx: Option<Receiver<Vec<(TopicFilter, QualityOfService)>>> nw_request_tx: Option<Sender<NetworkRequest>> nw_notification_rx: Option<Receiver<NetworkNotification>> misc_rx: Option<Receiver<MiscNwRequest>> message_callback: Option<Arc<Box<Fn(Message) + Send + Sync>>>

On message callback

publish_callback: Option<Arc<Box<Fn(Message) + Send + Sync>>>

On publish callback

poll: Poll

Poll

Methods

impl MqttClient

fn new(opts: MqttOptions) -> Self

fn mock_start(self) -> Result<(Self, MqRequest, Receiver<NetworkRequest>)>

fn start(self) -> Result<MqRequest>

Connects to the broker and starts an event loop in a new thread. Returns 'Request' and handles reqests from it. Also handles network events, reconnections and retransmissions.

fn message_callback<F>(self, callback: F) -> Self where F: Fn(Message) + Send + Sync + 'static

Set the message callback. This is called when a message is received from the broker.

fn publish_callback<F>(self, callback: F) -> Self where F: Fn(Message) + Send + Sync + 'static

Set the publish callback. This is called when a published message has been sent to the broker successfully.