Struct rumqtt::MqttOptions

pub struct MqttOptions {
    pub addr: String,
    pub keep_alive: Option<u16>,
    pub clean_session: bool,
    pub client_id: Option<String>,
    pub username: Option<String>,
    pub password: Option<String>,
    pub reconnect: Option<u16>,
    pub will: Option<(String, String)>,
    pub will_qos: QualityOfService,
    pub will_retain: bool,
    pub pub_q_len: u16,
    pub sub_q_len: u16,
    pub queue_timeout: u16,
    pub ca: Option<PathBuf>,
    pub verify_ca: bool,
    pub client_cert: Option<(PathBuf, PathBuf)>,
}

Fields
addr: String keep_alive: Option<u16> clean_session: bool client_id: Option<String> username: Option<String> password: Option<String> reconnect: Option<u16> will: Option<(String, String)> will_qos: QualityOfService will_retain: bool pub_q_len: u16 sub_q_len: u16 queue_timeout: u16 ca: Option<PathBuf> verify_ca: bool client_cert: Option<(PathBuf, PathBuf)>
Methods

impl MqttOptions

fn new() -> MqttOptions

Creates a new MqttOptions object which is used to set connection options for new client. Below are defaults with which this object is created.

client_id 	Randomly generated

clean_session 	true

keep_alive 	10 secs

reconnect try 	Doesn't try to reconnect

retransmit try time 	60 secs

pub_q_len 	50

sub_q_len 	5

fn set_keep_alive(self, secs: u16) -> Self

Number of seconds after which client should ping the broker if there is no other data exchange

fn set_client_id(self, client_id: &str) -> Self

Client id of the client. A random client id will be selected if you don't set one

fn set_clean_session(self, clean_session: bool) -> Self

clean_session = true instructs the broker to clean all the client state when it disconnects. Note that it is broker which is discarding the client state. But this client will hold its queues and attemts to to retransmit when reconnection happens. (TODO: Verify this)

When set false, broker will hold the client state and performs pending operations on the client when reconnection with same client_id happens.

Hence make sure that you manually set client_id when clean_session is false

fn set_user_name(self, username: &str) -> Self

Set username for broker to perform client authentication via username and password

fn set_password(self, password: &str) -> Self

Set password for broker to perform client authentication vis username and password

fn set_pub_q_len(self, len: u16) -> Self

All the QoS > 0 publishes state will be saved to attempt retransmits incase ack from broker fails.

If broker disconnects for some time, Publisher shouldn't throw error immediately during publishes. At the same time, Publisher shouldn't be allowed to infinitely push to the queue.

Publish queue length specifies maximum queue capacity upto which Publisher can push with out blocking. Messages in this queue will published as soon as connection is reestablished and Publisher gets unblocked

fn set_sub_q_len(self, len: u16) -> Self

fn set_q_timeout(self, secs: u16) -> Self

fn set_reconnect(self, dur: u16) -> Self

Time interval after which client should retry for new connection if there are any disconnections. By default, no retry will happen

fn set_will(self, will_topic: &str, will_message: &str) -> Self

Set will for the client so that broker can send will_message on will_topic when this client ungracefully dies.

fn set_will_qos(self, qos: QualityOfService) -> Self

Set QoS for the will message

fn set_will_retain(self, retain: bool) -> Self

Set will retian so that future clients subscribing to will topic knows of client's death.

fn set_ca<P>(self, cafile: P) -> Self where P: AsRef<Path>

Set CA file for server authentication during TLS connection

fn set_should_verify_ca(self, should_verify_ca: bool) -> Self

Set flag to determine whether or not to verify server CA during TLS connection

fn set_client_cert<P>(self, certfile: P, keyfile: P) -> Self where P: AsRef<Path>

Set client cert and key for server to do client authentication

fn broker(self, addr: &str) -> Self

Creates a new mqtt client with the broker address that you want to connect to. Along with connection details, this object holds all the state information of a connection.

NOTE: This should be the final call of MqttOptions method chaining

let client = client_options.set_keep_alive(5)
                          .set_reconnect(5)
                          .set_client_id("my-client-id")
                          .set_clean_session(true)
                          .connect("localhost:1883");Run