Struct rumqtt::MqttClient

pub struct MqttClient {
    pub addr: SocketAddr,
    pub opts: MqttOptions,
    pub stream: NetworkStream,
    pub state: MqttState,
    pub last_flush: Instant,
    pub last_pkid: PacketIdentifier,
    pub await_ping: bool,
    pub initial_connect: bool,
    pub pub0_channel_pending: u32,
    pub pub1_channel_pending: u32,
    pub pub2_channel_pending: u32,
    pub disconnect_block: bool,
    pub should_qos1_block: bool,
    pub should_qos2_block: bool,
    pub no_of_reconnections: u32,
    pub pub0_rx: Option<Receiver<Message>>,
    pub pub1_rx: Option<Receiver<Message>>,
    pub pub2_rx: Option<Receiver<Message>>,
    pub sub_rx: Option<Receiver<Vec<(TopicFilter, QualityOfService)>>>,
    pub incoming_rx: Option<Receiver<VariablePacket>>,
    pub mionotify_tx: Option<Sender<MioNotification>>,
    pub connsync_tx: Option<SyncSender<MqttStatus>>,
    pub streamupdate_tx: Option<SyncSender<NetworkStream>>,
    pub outgoing_pub: VecDeque<(i64, Box<Message>)>,
    pub incoming_rec: VecDeque<Box<Message>>,
    pub outgoing_rec: VecDeque<(i64, Box<Message>)>,
    pub outgoing_rel: VecDeque<(i64, PacketIdentifier)>,
    pub subscriptions: VecDeque<Vec<(TopicFilter, QualityOfService)>>,
    pub message_callback: Option<Arc<MessageSendableFn>>,
    pub publish_callback: Option<Arc<PublishSendableFn>>,
    pub pool: Option<ThreadPool>,
}

Handles commands from Publisher and Subscriber. Saves MQTT state and takes care of retransmissions.

Fields
addr: SocketAddr opts: MqttOptions stream: NetworkStream state: MqttState last_flush: Instant last_pkid: PacketIdentifier await_ping: bool initial_connect: bool pub0_channel_pending: u32 pub1_channel_pending: u32 pub2_channel_pending: u32 disconnect_block: bool should_qos1_block: bool should_qos2_block: bool no_of_reconnections: u32 pub0_rx: Option<Receiver<Message>> pub1_rx: Option<Receiver<Message>> pub2_rx: Option<Receiver<Message>> sub_rx: Option<Receiver<Vec<(TopicFilter, QualityOfService)>>> incoming_rx: Option<Receiver<VariablePacket>> mionotify_tx: Option<Sender<MioNotification>> connsync_tx: Option<SyncSender<MqttStatus>> streamupdate_tx: Option<SyncSender<NetworkStream>> outgoing_pub: VecDeque<(i64, Box<Message>)>

Queues. Note: 'record' is qos2 term for 'publish' For QoS 1. Stores outgoing publishes

incoming_rec: VecDeque<Box<Message>>

For QoS 2. Store for incoming publishes to record.

outgoing_rec: VecDeque<(i64, Box<Message>)>

For QoS 2. Store for outgoing publishes.

outgoing_rel: VecDeque<(i64, PacketIdentifier)>

For Qos2. Store for outgoing pubrel packets.

subscriptions: VecDeque<Vec<(TopicFilter, QualityOfService)>> message_callback: Option<Arc<MessageSendableFn>>

On message callback

publish_callback: Option<Arc<PublishSendableFn>>

On publish callback

pool: Option<ThreadPool>
Methods

impl MqttClient

fn new(opts: MqttOptions) -> Self

fn message_callback<F>(self, callback: F) -> Self where F: Fn(Message) + Send + Sync + 'static

Set the message callback. This is called when a message is received from the broker.

fn publish_callback<F>(self, callback: F) -> Self where F: Fn(u16) + Send + Sync + 'static

Set the publish callback. This is called when a published message has been sent to the broker successfully.

fn start(self) -> Result<(Publisher, Subscriber)>

Connects to the broker and starts an event loop in a new thread. Returns Subscriber and Publisher and handles reqests from them. Also handles network events, reconnections and retransmissions.

fn get_reconnection_count(&self) -> u32

Return a count of (successful) mqtt connections that happened from the start. Just to know how many times the client reconnected (coz of bad networks, broker crashes etc)

fn _disconnect(&mut self) -> Result<()>