Struct rumqtt::MqttClient

pub struct MqttClient {
    pub addr: SocketAddr,
    pub opts: MqttOptions,
    pub stream: NetworkStream,
    pub state: MqttState,
    pub last_flush: Instant,
    pub last_pkid: PacketIdentifier,
    pub await_ping: bool,
    pub initial_connect: bool,
    pub pub1_channel_pending: u32,
    pub pub2_channel_pending: u32,
    pub should_qos1_block: bool,
    pub should_qos2_block: bool,
    pub no_of_reconnections: u32,
    pub pub0_rx: Option<Receiver<Message>>,
    pub pub1_rx: Option<Receiver<Message>>,
    pub pub2_rx: Option<Receiver<Message>>,
    pub sub_rx: Option<Receiver<Vec<(TopicFilter, QualityOfService)>>>,
    pub connsync_tx: Option<SyncSender<MqttStatus>>,
    pub mionotify_tx: Option<Sender<MioNotification>>,
    pub outgoing_pub: VecDeque<(i64, Box<Message>)>,
    pub incoming_rec: VecDeque<Box<Message>>,
    pub outgoing_rec: VecDeque<(i64, Box<Message>)>,
    pub outgoing_rel: VecDeque<(i64, PacketIdentifier)>,
    pub subscriptions: VecDeque<Vec<(TopicFilter, QualityOfService)>>,
    pub callback: Option<Arc<SendableFn>>,
}

Handles commands from Publisher and Subscriber. Saves MQTT state and takes care of retransmissions.

Fields

addr

opts

stream

state

last_flush

last_pkid

await_ping

initial_connect

pub1_channel_pending

pub2_channel_pending

should_qos1_block

should_qos2_block

no_of_reconnections

pub0_rx

pub1_rx

pub2_rx

sub_rx

connsync_tx

mionotify_tx

outgoing_pub

Queues. Note: 'record' is qos2 term for 'publish' For QoS 1. Stores outgoing publishes

incoming_rec

For QoS 2. Store for incoming publishes to record.

outgoing_rec

For QoS 2. Store for outgoing publishes.

outgoing_rel

For Qos2. Store for outgoing pubrel packets.

subscriptions

callback

On message callback

Methods

impl MqttClient

fn new(opts: MqttOptions) -> Self

fn message_callback<F>(self, callback: F) -> Self where F: Fn(Message) + Send + Sync + 'static

fn start(self) -> Result<(Publisher, Subscriber)>

Connects to the broker and starts an event loop in a new thread. Returns Subscriber and Publisher and handles reqests from them. Also handles network events, reconnections and retransmissions.

fn get_reconnection_count(&self) -> u32

Return a count of (successful) mqtt connections that happened from the start. Just to know how many times the client reconnected (coz of bad networks, broker crashes etc)

fn _disconnect(&mut self) -> Result<()>